+++ examples/miner_commit/src/app.py
@@ -3,7 +3,7 @@ import logging
 import pathlib
 
 from fastapi import FastAPI, Body, HTTPException
-from data_types import MinerInput, MinerOutput
+from data_types import CommitFile, MinerInput, MinerOutput
 
 logger = logging.getLogger(__name__)
 logging.basicConfig(
@@ -31,13 +31,18 @@ def solve(miner_input: MinerInput = Body(...)) -> MinerOutput:
         _src_dir = pathlib.Path(__file__).parent.resolve()
         _commit_dir = _src_dir / "commit"
         with open(_commit_dir / "train.py") as _training_file:
-            _training_script = _training_file.read()
+            _training_content = _training_file.read()
         with open(_commit_dir / "submissions.py") as _submission_file:
-            _inference_script = _submission_file.read()
+            _inference_content = _submission_file.read()
 
         _miner_output = MinerOutput(
-            train_script=_training_script,
-            inference_script=_inference_script,
+            commit_files=[
+                CommitFile(file_name="train.py", content=_training_content),
+                CommitFile(
+                    file_name="submissions.py",
+                    content=_inference_content,
+                ),
+            ]
         )
         logger.info("Successfully retrieved commit files.")
     except Exception as err: